Migration from EdgeDB to PostgreSQL
It was fun to have EdgeDB, but it's just not there yet. Query language
is nice, but it has steep learning curve. But problem is deployment.
There are no GUIs that would be usable and CLI is just rough. And also
the Docker container uses a LOT of memory, so I ended up just rewriting
it. I wanted to try Drizzle so here we are.